+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0](https://github.com/Bankole2000/cp-frontend/compare/v1.0.1...v1.1.0) (2023-01-30)
+
+
+### Features
+
+* **listings:** admin can crud listing types ([b99ef64](https://github.com/Bankole2000/cp-frontend/commits/b99ef6456a5080d5e790a3c83e2ff8ff1a36968b))
+
+
+* **images:** change image sizes ([a7f127e](https://github.com/Bankole2000/cp-frontend/commits/a7f127e22ec3eaf13edc6ae7defd4573a5251714))
+
 ### [1.0.1](https://github.com/Bankole2000/cp-frontend/compare/v1.0.0...v1.0.1) (2023-01-18)